Documentation + Tweaks
added in code documentation for every function, removed the unnecessary boolean variables for the four passive item effects, tweaked heal bonuses: heal bonuses on an item can now both be a fixed sum and a multiplier together, and the overall process of applying a heal bonus calculates the multiplier and flat bonus for all items combined and applies that to the heal bonus, meaning that healing will be the same regardless of equipment order. Additionally, altered a few getter functions that used variables where they should have been using getters.