File structure change + lib.lua
The Pie repository now has the library at its root, rather than root/pie/. The pie_examples library is also no longer part of the repository, however up-to-date copies of the examples library will instead be provided with each release.
These changes have been made as it seems more intuitive to have the library at the root than as part of a subfolder.
Also added lib.lua and added a new load message that is printed to the console.